fix xstatus

+++ Sawim/src/main/java/protocol/Profile.java
@@ -9,7 +9,7 @@ public final class Profile {
     public byte statusIndex = StatusInfo.STATUS_OFFLINE;
     public String statusMessage;
 
-    public byte xstatusIndex = -1;
+    public int xstatusIndex = -1;
     public String xstatusTitle;
     public String xstatusDescription;
     public boolean isActive;
